created demo projects
