+++ deploy/demo/kafka-client-pod.yaml
@@ -1,0 +1,16 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: kafka-client
+  namespace: kafscale-demo
+  labels:
+    app: kafka-client
+spec:
+  containers:
+  - name: kafka-client
+    image: apache/kafka:3.9.1
+    command: ["/bin/bash", "-c", "sleep infinity"]
+    resources:
+      limits:
+        memory: "512Mi"
+        cpu: "500m"
