+++ deploy/demo/kafscale-cluster.yaml
@@ -1,0 +1,33 @@
+# Copyright 2025 Alexander Alten (novatechflow), NovaTechflow (novatechflow.com).
+# This project is supported and financed by Scalytics, Inc. (www.scalytics.io).
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: kafscale.io/v1alpha1
+kind: KafscaleCluster
+metadata:
+  name: kafscale
+  namespace: kafscale-demo
+spec:
+  brokers:
+    image:
+      repository: ghcr.io/novatechflow/kafscale-broker
+      tag: ""
+    replicas: 1
+  s3:
+    bucket: kafscale-snapshots
+    region: us-east-1
+    endpoint: http://minio.kafscale-demo.svc.cluster.local:9000
+    credentialsSecretRef: kafscale-s3-credentials
+  etcd:
+    endpoints: []
