+++ deploy/demo/kafscale-topics.yaml
@@ -1,0 +1,46 @@
+# Copyright 2025 Alexander Alten (novatechflow), NovaTechflow (novatechflow.com).
+# This project is supported and financed by Scalytics, Inc. (www.scalytics.io).
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: kafscale.io/v1alpha1
+kind: KafscaleTopic
+metadata:
+  name: demo-topic-1
+  namespace: kafscale-demo
+spec:
+  clusterRef: kafscale
+  partitions: 1
+---
+apiVersion: kafscale.io/v1alpha1
+kind: KafscaleTopic
+metadata:
+  name: demo-topic-2
+  namespace: kafscale-demo
+spec:
+  clusterRef: kafscale
+  partitions: 1
+---
+apiVersion: kafscale.io/v1alpha1
+kind: KafscaleTopic
+metadata:
+  name: orders-springboot
+  namespace: kafscale-demo
+spec:
+  clusterRef: kafscale
+  partitions: 1
+
+
+
+
+
