--- docs/guide/examples/spring-boot-kafscale-demo/src/main/java/com/example/kafscale/KafScaleDemoApplication.java
@@ -1,14 +1,0 @@
-package com.example.kafscale;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.kafka.annotation.EnableKafka;
-
-@SpringBootApplication
-@EnableKafka
-public class KafScaleDemoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(KafScaleDemoApplication.class, args);
-    }
-}
