--- docs/guide/examples/spring-boot-kafscale-demo/src/main/java/com/example/kafscale/model/Order.java
@@ -1,72 +1,0 @@
-package com.example.kafscale.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class Order {
-
-    @JsonProperty("orderId")
-    private String orderId;
-
-    @JsonProperty("product")
-    private String product;
-
-    @JsonProperty("quantity")
-    private int quantity;
-
-    @JsonProperty("timestamp")
-    private long timestamp;
-
-    public Order() {
-        this.timestamp = System.currentTimeMillis();
-    }
-
-    public Order(String orderId, String product, int quantity) {
-        this.orderId = orderId;
-        this.product = product;
-        this.quantity = quantity;
-        this.timestamp = System.currentTimeMillis();
-    }
-
-    // Getters and Setters
-    public String getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(String orderId) {
-        this.orderId = orderId;
-    }
-
-    public String getProduct() {
-        return product;
-    }
-
-    public void setProduct(String product) {
-        this.product = product;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    public long getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(long timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    @Override
-    public String toString() {
-        return "Order{" +
-                "orderId='" + orderId + '\'' +
-                ", product='" + product + '\'' +
-                ", quantity=" + quantity +
-                ", timestamp=" + timestamp +
-                '}';
-    }
-}
