+++ examples/E20_spring-boot-kafscale-demo/Dockerfile
