+++ examples/E20_spring-boot-kafscale-demo/src/main/java/com/example/kafscale/KafScaleDemoApplication.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2025 Alexander Alten (novatechflow), NovaTechflow (novatechflow.com).
+ * This project is supported and financed by Scalytics, Inc. (www.scalytics.io).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.kafscale;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.kafka.annotation.EnableKafka;
+
+@SpringBootApplication
+@EnableKafka
+public class KafScaleDemoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(KafScaleDemoApplication.class, args);
+    }
+}
