+++ examples/E20_spring-boot-kafscale-demo/src/main/java/com/example/kafscale/model/Order.java
@@ -1,0 +1,89 @@
+/*
+ * Copyright 2025 Alexander Alten (novatechflow), NovaTechflow (novatechflow.com).
+ * This project is supported and financed by Scalytics, Inc. (www.scalytics.io).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.kafscale.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class Order {
+
+    @JsonProperty("orderId")
+    private String orderId;
+
+    @JsonProperty("product")
+    private String product;
+
+    @JsonProperty("quantity")
+    private int quantity;
+
+    @JsonProperty("timestamp")
+    private long timestamp;
+
+    public Order() {
+        this.timestamp = System.currentTimeMillis();
+    }
+
+    public Order(String orderId, String product, int quantity) {
+        this.orderId = orderId;
+        this.product = product;
+        this.quantity = quantity;
+        this.timestamp = System.currentTimeMillis();
+    }
+
+    // Getters and Setters
+    public String getOrderId() {
+        return orderId;
+    }
+
+    public void setOrderId(String orderId) {
+        this.orderId = orderId;
+    }
+
+    public String getProduct() {
+        return product;
+    }
+
+    public void setProduct(String product) {
+        this.product = product;
+    }
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+
+    public long getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(long timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    @Override
+    public String toString() {
+        return "Order{" +
+                "orderId='" + orderId + '\'' +
+                ", product='" + product + '\'' +
+                ", quantity=" + quantity +
+                ", timestamp=" + timestamp +
+                '}';
+    }
+}
