+++ examples/E20_spring-boot-kafscale-demo/src/main/resources/application.yml
@@ -1,0 +1,55 @@
+# Copyright 2025 Alexander Alten (novatechflow), NovaTechflow (novatechflow.com).
+# This project is supported and financed by Scalytics, Inc. (www.scalytics.io).
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+server:
+  port: 8083
+
+logging:
+  level:
+    root: INFO
+    org.springframework.kafka: INFO
+    org.apache.kafka: INFO
+
+spring:
+  kafka:
+    bootstrap-servers: 127.0.0.1:39092
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+      acks: 0
+      retries: 3
+      properties:
+        linger.ms: 10
+        batch.size: 16384
+        enable.idempotence: false # false for KafScale
+    consumer:
+      group-id: kafscale-demo-group-${random.uuid}
+      auto-offset-reset: earliest
+      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      properties:
+        spring.json.trusted.packages: "*"
+      enable-auto-commit: true
+      auto-commit-interval: 1
+    listener:
+      ack-mode: batch
+      concurrency: 1
+    admin:
+      properties:
+        bootstrap.servers: 127.0.0.1:39092
+
+app:
+  kafka:
+    topic: orders-springboot
