+++ examples/E20_spring-boot-kafscale-demo/src/main/resources/static/index.html
